[plugin/auth_ldap] PS-7352, PS-7353: Fix LDAP issues on reconfiguration
Issues:
* Specifying a non existent server with to the LDAP variables could
  result in a slow timeout, making the SET command slow to return,
  blocking connections in the meantime.
* Specifying a non existent CA file for TLS/SSL when there are already
  active connections will result in a crash in the ldap library

Fix: make the variables that requires reconnecting to a new server static,
so no runtime reconfiguration is required. These parameters only have to
be changed when the LDAP server is moved, which is not a common event.

------------------------------------------------------------------------

PS-7358: Bypass password verification with the LDAP plugin

Issue: Active Directory treats empty passwords as anonymous login, and
returns a success status. This results in the simple_ldap plugin logging
in users when they didn't provide any passwords.

Fix: explicitly disable empty passwords in the ldap auth plugin.