PS-8626: Inserting JSON data through LOAD DATA INFILE doesn't show failed record.
When LOAD DATA or multi-row INSERT statement tried to store bad data in a
JSON column an error message was emitted, which didn't reference number
of row which caused the problem. Such behavior was not consistent with
how similar problems were reported for columns of other types.
It didn't allow user to understand which record among inserted
was problematic.

This patch solves the problem by changing message for ER_INVALID_JSON_TEXT
error, which is emitted in such cases, to mention number of row being
inserted. It is easy to do so, as the information about row being processed
is already available at the point where error is emitted from
Diagnostics_area::current_row_for_condition() member function.

It makes behavior for JSON type consistent with how warnings and errors
are reported for other types.