[plugin/clone] PS-8188 : Make the clone_plugin to not force plugins to match on recipient/donor (#5032)
PS-8188 : Make the clone_plugin to not force plugins to match on recipient/donor

https://jira.percona.com/browse/PS-8188

Problem:
--------
Clone enforces donor and recipient have same plugins installed.
For example, if a donor has an extra authentication plugin or a data masking plugin
installed but not installed on recipient, it is not possible to clone.
It will fail with the below error.

'Clone Donor plugin <plugin_name> is not active in Recipient.'

Fix:
----
Introduce new clone option 'clone_exclude_plugins_match' to take list of plugins that
are not installed on recipient.

This option is used on recipient server. The following plugins are disallowed. They must
be installed on recipient.

"daemon_keyring_proxy_plugin",
"binlog",
"performance_schema",
"memory",
"innodb",
"keyring_file",
"keyring_vault"

Co-authored-by: Satya Bodapati <satya.bodapati@percona.com>

--------------------------------------------------------------------------

PS-8844: Fix the failing clone.plugin_mismatch

https://jira.percona.com/browse/PS-8844

The test clone.plugin_mismatch fails when run from installed
directories due to the different path of the plugin directory used by
the test.

Failure:
        CURRENT_TEST: clone.plugin_mismatch
        mysqltest: At line 28: Command "remove_file" failed with error 1. my_errno=2.

The test involves simulating the secnario of missing plugins on
recipient server. To do the same, the test earlier used to copy all
plugins to recipient server's plugin_dir and delete some plugins to test
the clone plugin's behavior when recipeint had some plugins missing.

During the test investigation, it was found that it is just sufficient
to have only the clone plugin copied to the recipient server so that it
simulates the missing plugins.

This patch makes the the test independent of plugin paths so it works on
both build and installed directories.