[upstream] PS-8871 : RESOURCE GROUP hint fails when used within a prepared statement
https://jira.percona.com/browse/PS-8871

RESOURCE_GROUP hints specified in the text of prepared statement didn't
apply during its execution.

The problem occurred because information about RESOURCE_GROUP hint (the
resource group to be used for statement execution) was associated with
connection context (a.k.a. THD) and not with prepared statement context.
Moreover, to avoid situation when hint specified for a statement affected
execution of further statements in connection, it was reset after execution
of the statement. Which is not re-execution safe technique.

This patch solves the problem by associating information about this hint
with LEX object which is part of prepared statement context. The problem
with a hint specified for a statement affecting execution of further
statements in the connection is solved automatically in this case, since
LEX is re-inited before parsing each statement.

----------------------------------------------------------------------

PS-9328: Merge 8.4.2 - post merge MTR fixes.

Updated results for Percona-only part of main.resource_group_bugs test
after Upstream's fix for bug#33057164 "PSI_THREAD_INFO not updated while
executing PREPARED STATEMENTS".

https://github.com/mysql/mysql-server/commit/987fc9a5ea65e0889a1a60ab4a68f2db1b038ac7