set up tighter iterator bounds in myrocks
Upstream commit ID: https://github.com/facebook/mysql-5.6/commit/87aa2c2951fc
PS-9395: Merge percona-202404 (https://jira.percona.com/browse/PS-9395)

Summary:
MyRocks supports per-row TTL for tables, with some common config at the table level, for example:

  CREATE TABLE t1 (
    a int NOT NULL,
    b int NOT NULL,
    c int NOT NULL,
    ts bigint UNSIGNED NOT NULL,
    PRIMARY KEY (a),
    KEY kb (b) COMMENT 'kb',
    KEY kc (c) COMMENT 'kc'
  ) ENGINE=ROCKSDB
  COMMENT='ttl_duration=1;ttl_col=ts;';
where individual rows expire after `ttl_duration` seconds have passed (1) from either the time when the rows were created or updated (implicit timestamp), or (2) from the timestamp value in the column specified as `ttl_col=ts` in the above case (explicit timestamp). In the second scheme, the `ttl_col` value can be updated by the user.

If a lot of rows are created with an implicit timestamp scheme, or if the same ttl_col value is used for a lot of rows, then it implies that a lot of rows will expire roughly around the same time. Till the next compaction job runs, when these rows get "taken care of", any range scan queries running on such tables may end up having to deal with a lot of expired rows that should be skipped (TTL filtering) before sending responses back to the user.

TTL filtering is implemented in the MyRocks layer within the RDB iterators that directly call the RocksDB iterators. If the index range scan AccessMethod is being used by the SQL layer, the index conditions get pushed down to the storage engine (ICP). The range conditions are checked in MyRocks at the handler API level (outer) iterators, which in turn instantiate and call RDB iterators.

If the range scan encounters a large number of consecutive expired rows, then this can lead to the RDB iterator continuously filtering all these rows in a long running loop. In fact, the RDB and RocksDB iterators may have long gone past the actual range conditions. This can lead to long running queries that would have returned very quickly if there was a way to check range conditions in the outer iterator from time to time, or if the iterator bounds were tight enough that RocksDB itself would detect the end of the iterator range and return much sooner.

The iterator bounds currently being set by MyRocks for RocksDB iterators was very conservative (read: loose). It would check for the common prefix for the lower and upper bounds and, using that as a starting point, would use the predecessor and successor for this common prefix as the iterator bound. Because of this, the "valid range" for a RockDB iterator was much wider than what was dictated by the SQL query. This would normally not cause any issues as the query iterators would always seek to the start range, and the outer level MyRocks iterators will end the RocksDB iterator as soon as the end range is reached. If we hit the scenario detailed above, because of the TTL check implemented in the lower level RDBIterator, the outer MyRocks iterators don't get a chance to perform their end range check, while the loose iterator bounds at the RockDB level means that the RocksDB level iterator keeps on iterating way past the SQL query upper bound.

This fix could also be implemented by allowing the TTL check to be deferred till after the range check if we fail a configured # of TTL checks. This solution works as well, and is less risky overall: D54104480.

A better fix is to tighten up the iterator bounds themselves. This has the advantage of benefiting all queries (not just the ones affected by the TTLDB issue), as well as keeping the TTL checking logic in one place (easier to reason about). This is the approach taken in the present diff, and can be more risky as, again, it would affect all queries. A large number of MTR tests have been added to test the correctness of the boundary conditions in several ways.

Differential Revision: D55305366

fbshipit-source-id: 4a17514a25633b0311fd8cf74d7644fc80fe30cc