[upstream] PS-9703 "Upstream 8.0.41 release does not fully fix PS-9144"
https://perconadev.atlassian.net/browse/PS-9703

Problem:
--------
ALTER TABLE which rebuilds InnoDB table using INPLACE algorithm might sometimes
lead to row loss if concurrent purge happens on the table being ALTERed.

Analysis:
---------
This issue was introduced in Upstream version 8.0.41 as unwanted side-effect
of fixes for bug#115608 (PS-9144), in which similar problem is observed but
in a different scenario, and bug#115511 (PS-9214). It was propageted to
Percona Server 8.0.41-32, in which we opted for reverting our versions
of fixes for PS-9144 and PS-9214 in favour of Upstream ones.

New implementation of parallel ALTER TABLE INPLACE in InnoDB was introduced in
MySQL 8.0.27. Its code is used for online table rebuild even in a single-thread
case.
This implementation iterates over all the rows in the table, in general case,
handling different subtrees of a B-tree in different threads. This iteration
over table rows needs to be paused, from time to time, to commit InnoDB MTR/
release page latches it holds. This is necessary to give a way to concurrent
actions on the B-tree scanned (this happens when switching to the next page)
or before flushing rows of new version of table from in-memory buffer to the
B-tree. In order to resume iteration after such pause persistent cursor
position saved before pause is restored.

The problem described above occurs when we try to save and then restore
position of cursor pointing to page supremum, before switching to the next
page. In post-8.0.41 code this is done by simply calling
btr_pcur_t::store_position()/restore_position() methods for cursor that
point to supremum. In 8.0.42-based code this is done in
PCursor::save_previous_user_record_as_last_processed() and
PCursor::restore_to_first_unprocessed() pair of methods.

However, this doesn't work correctly in scenario, when after we have
saved cursor position and then committed mini-transaction/released latches
on the current page the next page is merged into the current one (after
purge removes some records from it). In this case the cursor position is
still restored as pointing to page supremum, and thus rows which were
moved over by merge are erroneously skipped.

***

Let us take look at an example. Let us assume that we have two pages
p1 : [inf, a, b, c, sup] and the next one p2 : [inf, d, e, f, sup].
Our thread which is one of the parallel ALTER TABLE worker threads
has completed scan of p1, so its cursor positioned on p1:'sup' record.
Now it needs to switch to page p2, but also give a way to threads
concurrently updating the table. So it needs to make cursor savepoint,
commit mini-transaction and release the latches.

In post-8.0.41 code we simply do btr_pcur_t::store_position()/
restore_position() with the cursor positioned on p1 : 'sup' record,
then the following might happen: concurrent purge on page p2 might
delete some record from it (e.g. 'f') and decide to merge of this page
into the page p1.
If this happens while latches are released this merge would go through
and and resulting in page p1 with the following contents
p1 : [inf, a, b, c, d, e, sup]. Savepoint for p1 : 'sup' won't
be invalidated (one can say that savepoints for sup and inf are not
safe against concurrent merges in this respect) and after restoration
of cursor the iteration will continue, on the next page, skipping
records 'd' and 'e'.

***

Fix:
----
This patch solves the problem by working around the issue with saving/
restoring cursor pointing to supremum. Instead of storing position of
supremum record PCursor::save_previous_user_record_as_last_processed()
now stores the position of record that precedes it.
And then PCursor::restore_to_first_unprocessed() does restore in two
steps - 1) restores position of this preceding record (or its closest
precedessor if it was purged meanwhile) and then 2) moves one step
forward assuming that will get to the supremum record at which cursor
pointed originally. If this is not true, i.e. there is user record
added to the page by the merge (or simple concurrent insert), we
assume that this and following records are unprocessed. The caller
of PCursor::restore_to_first_unprocessed() detects this situation
by checking if cursor is positioned on supremum and handles by
resuming processing from record under the cursor if not.

***

Let us return to the above example to explain how fix works.
PCursor::save_previous_user_record_as_last_processed() does a step back
before calling btr_pcur_t::store_position(), so for cursor positioned
on p1 : 'sup' it is actually position corresponding to p1 : 'c' what
is saved. If the merge happens when latches are released, we still
get p1 : [inf, a, b, c, d, e, sup] and the savepoint is not invalidated.
PCursor::restore_to_first_unprocessed() calls btr_pcur_t::restore_position()
gets cursor pointing to p1 : 'c' as result, and then it tries to
compensate for step-back and moves cursor one step forward making it to
point to p1 : 'd'. Code which does  scanning detects the situation that
saving/restoring resulted in jump from supremum record to user record
and resume iteration  from p1 : 'd' without skipping any records.

***

Thanks to Bytedance team for bringing this issue to our attention!
The test case for this bug is based on the one that they reported
to the Upstream and later pointed to us.