PS-9220 [8.4]: Error during slow Query Log Rotation
https://perconadev.atlassian.net/browse/PS-9220

When multiple client threads are connected to the server, there can be
cases where the slow log purging after rotation attempt to delete the
older slow log files which don't exist anymore. In some of such cases,
the error returned by my_errno() after the unlink() method call was
inaccuarte causing error messages to be printed in the server error
log. So, my_errno() is replcaed with errno for accurate error returned.